+++ cfgov/processors/wordpress_faq.py
@@ -2,6 +2,7 @@ import sys
 import json
 import os.path
 import requests
+from sheerlike.external_links import process_external_links
 
 
 def posts_at_url(url):
@@ -45,6 +46,8 @@ def process_post(post):
                         post['tags'][i] = tag.replace(char, '')
 
     del post['custom_fields']
+
+    post = process_external_links(post)
 
     return {'_type': 'faq',
             '_id': post['slug'],