+++ CHANGELOG.md
@@ -19,7 +19,8 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 
 ### Changed
 
-- Abstracted create and queryOneout of the multiselect
+- Abstracted create and queryOneout of the multiselect.
+- Simplified array searching in the Multiselect.
 
 ### Removed